+++ src/main/java/com/travel/role/domain/user/dto/SignInRequestDTO.java
@@ -1,4 +1,4 @@
-package com.travel.role.global.auth.dto;
+package com.travel.role.domain.user.dto;
 
 import lombok.Data;
