+++ src/main/java/com/travel/role/global/auth/service/AuthService.java
@@ -17,8 +17,8 @@ import org.springframework.transaction.annotation.Transactional;
 import com.travel.role.domain.user.dao.UserRepository;
 import com.travel.role.domain.user.domain.UserEntity;
 import com.travel.role.global.auth.dto.AccessTokenResponse;
-import com.travel.role.global.auth.dto.SignInRequestDTO;
-import com.travel.role.global.auth.dto.SignUpRequestDTO;
+import com.travel.role.domain.user.dto.SignInRequestDTO;
+import com.travel.role.domain.user.dto.SignUpRequestDTO;
 import com.travel.role.global.auth.dto.TokenMapping;
 import com.travel.role.global.auth.exception.InvalidTokenException;
 import com.travel.role.global.auth.exception.NotExistTokenException;
