+++ src/test/java/com/travel/role/global/auth/service/AuthServiceTest.java
@@ -14,7 +14,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import com.travel.role.domain.user.dao.UserRepository;
 import com.travel.role.domain.user.domain.Role;
 import com.travel.role.domain.user.domain.UserEntity;
-import com.travel.role.global.auth.dto.SignUpRequestDTO;
+import com.travel.role.domain.user.dto.SignUpRequestDTO;
 import com.travel.role.global.auth.exception.auth.InvalidTokenException;
 import com.travel.role.global.auth.exception.auth.NotExistTokenException;
 import com.travel.role.global.exception.ExceptionMessage;
